Refactoring and moving dialog logic to a separate class.
